+++ corehq/apps/saved_reports/management/commands/find_saved_reports_with_slug.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, print_function
 from __future__ import unicode_literals
 from django.conf import settings
 from dimagi.utils.couch.cache import cache_core
-from corehq.apps.reports.models import ReportConfig
+from corehq.apps.saved_reports.models import ReportConfig
 
 from django.core.management import BaseCommand
