+++ corehq/apps/saved_reports/management/commands/test_reports.py
@@ -1,12 +1,12 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 from django.core.management.base import BaseCommand
-from corehq.apps.reports.tasks import daily_reports, weekly_reports
+from corehq.apps.saved_reports.tasks import daily_reports, weekly_reports
 
 
 class Command(BaseCommand):
     help = "Tests sending reports. Equvalent to firing the celery tasks right NOW."
-    
+
     def handle(self, **options):
         daily_reports()
         weekly_reports()
