+++ corehq/apps/saved_reports/tests/test_scheduled_reports.py
@@ -4,8 +4,8 @@ from __future__ import unicode_literals
 from datetime import datetime
 from django.test import SimpleTestCase, TestCase
 from corehq.apps.domain.shortcuts import create_domain
-from corehq.apps.reports.models import ReportNotification, ReportConfig
-from corehq.apps.reports.scheduled import guess_reporting_minute, get_scheduled_report_ids
+from corehq.apps.saved_reports.models import ReportConfig, ReportNotification
+from corehq.apps.saved_reports.scheduled import guess_reporting_minute, get_scheduled_report_ids
 from corehq.apps.reports.views import get_scheduled_report_response
 from corehq.apps.users.models import WebUser
