+++ corehq/apps/userreports/reports/view.py
@@ -35,7 +35,7 @@ from corehq.apps.locations.permissions import conditionally_location_safe
 from corehq.apps.reports.dispatcher import (
     ReportDispatcher,
 )
-from corehq.apps.reports.models import ReportConfig
+from corehq.apps.saved_reports.models import ReportConfig
 from corehq.apps.reports_core.exceptions import FilterException
 from corehq.apps.userreports.exceptions import (
     BadSpecError,
