+++ settings.py
@@ -299,6 +299,7 @@ HQ_APPS = (
     'corehq.messaging.smsbackends.airtel_tcl',
     'corehq.apps.reports.app_config.ReportsModule',
     'corehq.apps.reports_core',
+    'corehq.apps.saved_reports',
     'corehq.apps.userreports',
     'corehq.apps.aggregate_ucrs',
     'corehq.apps.data_interfaces',
@@ -1322,6 +1323,7 @@ COUCHDB_APPS = [
     'programs',
     'reminders',
     'reports',
+    'saved_reports',
     'sms',
     'smsforms',
     'telerivet',
